Modified data structure following recent thoughts
Added a new column in main data array which is a standardized version
of the country named (no accent or anything else of the sort)
Hence, I replaced all the .png and .json by an "x" ; it means that there
is indeed data that needs to be recovered using the shortname in the
last column.
Also added dummy data for pinpoints.